agentic_ai_system/pyproject.toml
[project]
name = "agentic_ai_system"
version = "0.1.0"
description = "Agentic AI System for Telangana Open Data Portal using CrewAI"
authors = [
    { name = "Neha Sureddi", email = "nehasureddi@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.177.0,<1.0.0",
    "crewai-tools",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0"
]

[project.scripts]
agentic_ai_system = "agentic_ai_system.main:run"
run_crew = "agentic_ai_system.main:run"
train = "agentic_ai_system.main:train"
replay = "agentic_ai_system.main:replay"
test = "agentic_ai_system.main:test"

[tool.poetry]
name = "agentic_ai_system"
version = "0.1.0"
description = "Agentic AI System for Telangana Open Data Portal using CrewAI"
authors = ["Neha Sureddi <nehasureddi@gmail.com>"]
readme = "README.md"
packages = [{ include = "agentic_ai_system", from = "src" }]
python = ">=3.10,<3.14"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
crewai = { extras = ["tools"], version = ">=0.177.0,<1.0.0" }
# I've commented out the original dependencies here since they are now in the [project] table

[tool.poetry.scripts]
agentic_ai_system = "agentic_ai_system.main:run"
run_crew = "agentic_ai_system.main:run"
train = "agentic_ai_system.main:train"
replay = "agentic_ai_system.main:replay"
test = "agentic_ai_system.main:test"

[tool.crewai]
type = "crew"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



